Feature/api auth

See merge request olly/CSC2022-Saggezza!23

+++ server/routes/project.py
@@ -5,17 +5,23 @@ from mongoengine import *
 from schema.project import *
 
 from model import Project
+from routes.auth import auth
 
 
 # Connect to mongodb
 connect('saggezza_db', host='localhost', port=27017)
 
 
+@auth.login_required
 class ProjectListAPI(Resource):
     # |- /project
     # |- POST: Create a new project
     # \- GET: Return all projects
     def post(self):
+        caller = get_caller(request)
+        if caller["role"] != "admin":
+            return res("⛔️ Only an admin can add a project", "error"), 401
+
         req = parse(request)
         errors = ProjectListSchema().validate(req)
         if errors:
@@ -26,7 +32,11 @@ class ProjectListAPI(Resource):
         project.save()
         return res('Project created successfully', 'success', project=convert_query(project))
 
+    @auth.login_required
     def get(self):
+        caller = get_bearer(request)
+        if caller["role"] != "admin":
+            return res("⛔️ Must be an admin to see all projects", "error"), 400
         categories = Project.objects().all()
         return res('All categories returned', 'success', categories=convert_query(categories))
 
@@ -37,7 +47,11 @@ class ProjectAPI(Resource):
     # |- GET: Return project
     # \- DELETE: Delete project
 
+    @auth.login_required
     def put(self, id):
+        caller = get_caller(request)
+        if caller["role"] != "admin":
+            return res("⛔️ Must be an admin to update a project", "error"), 400
         try:
             req = parse(request)
             errors = ProjectSchema().validate(req)
@@ -52,14 +66,22 @@ class ProjectAPI(Resource):
         except:
             return res("Project doesn't exist", 'error'), 400
 
+    @auth.login_required
     def get(self, id):
+        caller = get_bearer(request)
+        if caller["role"] != "manager":
+            return res("⛔️ Must be a manager to view a project", "error"), 400
         try:
             project = Project.objects(id=id)[0]
             return res('Retrieved Successfully', 'success', project=convert_query(project))
         except:
             return res("Project doesn't exist", 'error'), 400
 
+    @auth.login_required
     def delete(self, id):
+        caller = get_bearer(request)
+        if caller["role"] != "admin":
+            return res("⛔️ Must be an admin to delete a project", "error"), 400
         try:
             project = Project.objects(id=id)
             project.delete()
